added variable times to each setup step
